Update artifact name in GitHub workflow
The artifact name in the GitHub workflow for code quality analysis is updated to include the Python version. Previously, the Python-specific version of the 'coverage.xml' file was named simply 'coverage.xml'. Now, the Python version is clearly indicated in the name, making it easier to differentiate files when multiple versions of Python are tested.