Add skip_tests option to code_quality_analyzers workflow
The code_quality_analyzers workflow has been updated to include a new option 'skip_tests'. This option allows the user to skip the Python tests while mandatory to upload the coverage.xml artifact. This is useful for scenarios where we want to perform code quality analysis without running the tests.